+++ src/main/java/com/mapper/RulesMapper.java
@@ -31,7 +31,7 @@ public interface RulesMapper {
             @Result(property="faultId",column="fault_id",javaType=Integer.class),
             @Result(property="needPhenomenonId",column="need_phenomenon_id",javaType=Integer.class),
             @Result(property="needPhenomenonNum",column="need_phenomenon_num",javaType=Integer.class),
-            @Result(property="needPhenomenonThreshold",column="need_phenomenon_threshold",javaType=Integer.class),
+            @Result(property="needPhenomenonThreshold",column="need_phenomenon_threshold",javaType=Double.class),
     })
     public List<BeatsFault> getAllBeatsFault();
